requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
bidict==0.23.1
blinker==1.6.3
cachelib==0.13.0
cachetools==5.5.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.1.0
cmake==3.30.5
contourpy==1.1.1
cramjam==2.9.0
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
ecs-logging==2.1.0
einops==0.8.0
einx==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.4
fastparquet==2024.2.0
ffmpy==0.4.0
filelock==3.9.0
Flask==3.0.0
Flask-Session==0.8.0
Flask-SocketIO==5.4.1
flatbuffers==2.0.7
fonttools==4.54.1
frozendict==2.4.4
frozenlist==1.5.0
fsspec==2024.6.1
gast==0.4.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.23.4
idna==3.4
imageio==2.35.1
importlib-metadata==6.8.0
importlib_resources==6.4.5
influxdb==5.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
lightning-utilities==0.11.8
lit==18.1.8
loguru==0.7.2
lpips==0.1.4
lz4==4.3.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==0.6.1
msgspec==0.18.6
multidict==6.1.0
networkx==3.0
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.12.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.11
packaging==24.1
pandas==2.0.3
Pillow==9.3.0
piq==0.8.0
propcache==0.2.0
protobuf==3.20.1
psutil==5.9.5
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.8.2
python-engineio==4.10.1
python-multipart==0.0.17
python-socketio==5.11.4
pytorch-lightning==2.3.3
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
ruff==0.7.2
safetensors==0.4.3
scipy==1.10.1
semantic-version==2.10.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.12
taming-transformers==0.0.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorflow==2.7.1
tensorflow-estimator==2.7.0
tensorflow-gan==2.0.0
tensorflow-hub==0.9.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-probability==0.15.0
termcolor==2.4.0
timeout-decorator==0.5.0
timm==1.0.7
tomlkit==0.12.0
torch==2.0.0
torchaudio==2.0.0
torchmetrics==1.5.1
torchvision==0.15.1
tqdm==4.66.4
triton==2.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
vector_quantize_pytorch==1.14.8
websockets==12.0
Werkzeug==3.0.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.15.2
zipp==3.17.0
